(Unfinished) Added Popup scene and UI Fonts
Didn't get started on doing the scripts for this yet, but I've configured a simple UI layout and some fonts that we can use. Needs some way of setting the text when the scene is instantiated?

There's some methods to make the popup visible (which is a start i guess); https://docs.godotengine.org/en/3.1/classes/class_popup.html